+++ src/features/image.rs
@@ -8,6 +8,7 @@ use crate::Image;
 use crate::ImageInfo;
 use crate::ImageView;
 use crate::PixelFormat;
+use crate::error::UnsupportedImageFormat;
 
 impl AsImageView for image::DynamicImage {
 	fn as_image_view(&self) -> Result<ImageView, ImageDataError> {
@@ -36,7 +37,7 @@ impl From<image::DynamicImage> for Image {
 
 impl<P> AsImageView for image::ImageBuffer<P, Vec<u8>>
 where
-	P: image::Pixel<Subpixel = u8> + 'static,
+	P: image::Pixel<Subpixel = u8> + image::PixelWithColorType,
 {
 	fn as_image_view(&self) -> Result<ImageView, ImageDataError> {
 		let info = info(self)?;
@@ -47,7 +48,7 @@ where
 
 impl<P> AsImageView for &'_ image::ImageBuffer<P, Vec<u8>>
 where
-	P: image::Pixel<Subpixel = u8> + 'static,
+	P: image::Pixel<Subpixel = u8> + image::PixelWithColorType,
 {
 	fn as_image_view(&self) -> Result<ImageView, ImageDataError> {
 		(*self).as_image_view()
@@ -56,7 +57,7 @@ where
 
 impl<P> From<image::ImageBuffer<P, Vec<u8>>> for Image
 where
-	P: image::Pixel<Subpixel = u8> + 'static,
+	P: image::Pixel<Subpixel = u8> + image::PixelWithColorType,
 {
 	fn from(other: image::ImageBuffer<P, Vec<u8>>) -> Self {
 		let info = match info(&other) {
@@ -69,7 +70,10 @@ where
 }
 
 /// Consume an [`image::ImageBuffer`] and return the pixel data as boxed slice.
-fn into_bytes<P: 'static + image::Pixel<Subpixel = u8>>(buffer: image::ImageBuffer<P, Vec<u8>>) -> Box<[u8]> {
+fn into_bytes<P>(buffer: image::ImageBuffer<P, Vec<u8>>) -> Box<[u8]>
+where
+	P: image::Pixel<Subpixel = u8> + image::PixelWithColorType,
+{
 	buffer.into_raw().into_boxed_slice()
 }
 
@@ -83,13 +87,17 @@ fn dynamic_image_into_bytes(image: image::DynamicImage) -> Box<[u8]> {
 		image::DynamicImage::ImageRgba8(x) => into_bytes(x),
 		image::DynamicImage::ImageRgb16(_) => panic!("unsupported pixel format: Rgb16"),
 		image::DynamicImage::ImageRgba16(_) => panic!("unsupported pixel format: Rgba16"),
-		image::DynamicImage::ImageBgr8(x) => into_bytes(x),
-		image::DynamicImage::ImageBgra8(x) => into_bytes(x),
+		image::DynamicImage::ImageRgb32F(_) => panic!("unsupported pixel format: Rgb32F"),
+		image::DynamicImage::ImageRgba32F(_) => panic!("unsupported pixel format: Rgba32F"),
+		x => panic!("unsupported pixel format: {:?}", x),
 	}
 }
 
 /// Get the pixel data of an [`image::ImageBuffer`] to as a byte slice.
-fn as_bytes<P: 'static + image::Pixel<Subpixel = u8>>(buffer: &image::ImageBuffer<P, Vec<u8>>) -> &[u8] {
+fn as_bytes<P>(buffer: &image::ImageBuffer<P, Vec<u8>>) -> &[u8]
+where
+	P: image::Pixel<Subpixel = u8> + image::PixelWithColorType,
+{
 	&*buffer
 }
 
@@ -103,15 +111,16 @@ fn dynamic_image_as_bytes(image: &image::DynamicImage) -> &[u8] {
 		image::DynamicImage::ImageRgba8(x) => as_bytes(x),
 		image::DynamicImage::ImageRgb16(_) => panic!("unsupported pixel format: Rgb16"),
 		image::DynamicImage::ImageRgba16(_) => panic!("unsupported pixel format: Rgba16"),
-		image::DynamicImage::ImageBgr8(x) => as_bytes(x),
-		image::DynamicImage::ImageBgra8(x) => as_bytes(x),
+		image::DynamicImage::ImageRgb32F(_) => panic!("unsupported pixel format: Rgb32F"),
+		image::DynamicImage::ImageRgba32F(_) => panic!("unsupported pixel format: Rgba32F"),
+		x => panic!("unsupported pixel format: {:?}", x),
 	}
 }
 
 /// Extract the [`ImageInfo`] from an [`image::ImageBuffer`].
 fn info<P, C>(image: &image::ImageBuffer<P, C>) -> Result<ImageInfo, ImageDataError>
 where
-	P: image::Pixel<Subpixel = u8> + 'static,
+	P: image::Pixel<Subpixel = u8> + image::PixelWithColorType,
 	C: std::ops::Deref<Target = [u8]>,
 {
 	Ok(ImageInfo {
@@ -128,26 +137,19 @@ fn dynamic_image_info(image: &image::DynamicImage) -> Result<ImageInfo, ImageDat
 	match image {
 		image::DynamicImage::ImageLuma8(x) => info(x),
 		image::DynamicImage::ImageLumaA8(x) => info(x),
-		image::DynamicImage::ImageLuma16(_) => Err("unsupported pixel format: Luma16".into()),
-		image::DynamicImage::ImageLumaA16(_) => Err("unsupported pixel format: LumaA16".into()),
 		image::DynamicImage::ImageRgb8(x) => info(x),
 		image::DynamicImage::ImageRgba8(x) => info(x),
-		image::DynamicImage::ImageRgb16(_) => Err("unsupported pixel format: Rgb16".into()),
-		image::DynamicImage::ImageRgba16(_) => Err("unsupported pixel format: Rgba16".into()),
-		image::DynamicImage::ImageBgr8(x) => info(x),
-		image::DynamicImage::ImageBgra8(x) => info(x),
+		x => Err(UnsupportedImageFormat { format: format!("{:?}", x) }.into()),
 	}
 }
 
 /// Extract the PixelFormat from an [`image::Pixel`].
-fn pixel_format<P: image::Pixel>() -> Result<PixelFormat, ImageDataError> {
+fn pixel_format<P: image::PixelWithColorType>() -> Result<PixelFormat, ImageDataError> {
 	match P::COLOR_TYPE {
 		image::ColorType::L8 => Ok(PixelFormat::Mono8),
 		image::ColorType::La8 => Ok(PixelFormat::MonoAlpha8(Alpha::Unpremultiplied)),
-		image::ColorType::Bgr8 => Ok(PixelFormat::Bgr8),
-		image::ColorType::Bgra8 => Ok(PixelFormat::Bgra8(Alpha::Unpremultiplied)),
 		image::ColorType::Rgb8 => Ok(PixelFormat::Rgb8),
 		image::ColorType::Rgba8 => Ok(PixelFormat::Rgba8(Alpha::Unpremultiplied)),
-		x => Err(format!("unsupported color type: {:?}", x).into()),
+		x => Err(UnsupportedImageFormat { format: format!("{:?}", x) }.into()),
 	}
 }